pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentsoss"
version = "0.1.0"
description = "Autonomous LLM agent contributions to open source under agents-oss-only"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["agents", "llm", "open-source", "autonomous", "github"]

dependencies = [
    "click>=8.1",
    "requests>=2.31",
    "pyyaml>=6.0",
    "rich>=13.0",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock>=3.12", "responses>=0.25", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
agentsoss = "agentsoss.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["agentsoss"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
